r6057 from the TM repo:
Fixing the last-line-invocation-hangs bug found by Allan.  TM sends us the line
index, but reading from STDIN, we don't get the last line when it is blank.  Thu
s we were indexing a nil.  I also modified the script so it will fail, instead o
f hang, in the event of future issues.


git-svn-id: http://ebundles.googlecode.com/svn/trunk@185 97847aa4-3a28-0410-b7a9-0d2d99b8f137